BackEnd/ZeroOneTwo/requirements_ubunutu.txt
absl-py==0.9.0
appdirs==1.4.3
asgiref==3.2.7
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.12.48
botocore==1.15.48
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
filelock==3.0.12
gast==0.2.2
gensim==3.8.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
imutils==0.5.3
inflection==0.4.0
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
JPype1==0.7.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mecab-python===0.996-ko-0.9.2
mxnet==1.6.0
mxnet-cu101==1.6.0
mysqlclient==1.4.6
networkx==2.4
numpy==1.18.3
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pprint==0.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
smart-open==2.0.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
traitlets==4.3.3
tweepy==3.8.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
